Nest posts list request under term request
We need the term ID to request the posts list, so we need to send and receive the term query first
