+++ js/components/term/index.jsx
@@ -1,67 +1,61 @@
 /*global FoxhoundSettings */
 import React from 'react';
 import { connect } from 'react-redux';
+import DocumentMeta from 'react-document-meta';
+import BodyClass from 'react-body-class';
 
 // Internal dependencies
-import QueryPosts from 'wordpress-query-posts';
-import { isRequestingPostsForQuery, getPostsForQuery, getTotalPagesForQuery } from 'wordpress-query-posts/lib/selectors';
-
-// Components
-import TermHeader from './header';
-import PostList from 'components/posts/list';
-import Pagination from 'components/pagination/archive';
+import QueryTerm from 'wordpress-query-term';
+import { isRequestingTerm, getTermIdFromSlug, getTerm } from 'wordpress-query-term/lib/selectors';
 import Placeholder from 'components/placeholder';
+import List from './list';
 
-const Term = React.createClass( {
-	renderHeaderPlaceholder() {
-		return null;
-	},
+const TermHeader = ( { term, taxonomy, loading, termData = {}, query = {} } ) => {
+	const meta = {
+		title: termData.name + ' – ' + FoxhoundSettings.meta.title,
+		description: termData.description,
+	};
 
-	render() {
-		const posts = this.props.posts || [];
-
-		return (
-			<div className="card">
-				<TermHeader params={ { taxonomy: this.props.route.taxonomy, slug: this.props.params.slug } } />
-
-				<QueryPosts query={ this.props.query } />
-				{ this.props.loading ?
-					<Placeholder type="term" /> :
-					<PostList posts={ posts } />
-				}
-
-				<Pagination
-					path={ this.props.path }
-					current={ this.props.page }
-					isFirstPage={ 1 === this.props.page }
-					isLastPage={ this.props.totalPages === this.props.page } />
-			</div>
-		);
-	}
-} );
+	return (
+		<div className="card">
+			<DocumentMeta { ...meta } />
+			<BodyClass classes={ [ 'archive', taxonomy ] } />
+			<QueryTerm taxonomy={ taxonomy } termSlug={ term } />
+			{ loading ?
+				<Placeholder type="term" /> :
+				<div>
+					<header className="page-header">
+						<h1 className="page-title">{ termData.name }</h1>
+						{ termData.description && <p>{ termData.description }</p> }
+					</header>
+					<List query={ query } taxonomy={ taxonomy } term={ termData && termData.id } />
+				</div>
+			}
+		</div>
+	);
+}
 
 export default connect( ( state, ownProps ) => {
-	let query = {};
-	let path = FoxhoundSettings.URL.path || '/';
+	const term = ownProps.params.slug;
+	const taxonomy = ownProps.route.taxonomy;
+	const termId = getTermIdFromSlug( state, taxonomy, term );
+	const termData = getTerm( state, termId );
+	const requesting = isRequestingTerm( state, taxonomy, term );
+
+	const query = {};
 	query.paged = ownProps.params.paged || 1;
-	if ( 'category' === ownProps.route.taxonomy ) {
-		query.category_name = ownProps.params.slug;
-		path += 'category/' + ownProps.params.slug + '/';
+	if ( 'category' === taxonomy ) {
+		query.categories = [ termId ];
 	} else {
-		query.tag = ownProps.params.slug;
-		path += 'tag/' + ownProps.params.slug + '/';
+		query.tags = [ termId ];
 	}
 
-	const posts = getPostsForQuery( state, query );
-	const requesting = isRequestingPostsForQuery( state, query );
-
 	return {
-		path,
-		query,
-		posts,
+		term,
+		taxonomy,
+		termData,
 		requesting,
-		loading: requesting && ! posts,
-		page: parseInt( query.paged ),
-		totalPages: getTotalPagesForQuery( state, query ),
+		query,
+		loading: requesting && ! termData
 	};
-} )( Term );
+} )( TermHeader );
